+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- adds internationalize function
+
 # 0.6.1
 - Build binaries with node v16 -> works at runtime with node v8 -> v16 (and likely others)
 - Remove `-D_GLIBCXX_USE_CXX11_ABI=0` build flag